# docs/apis/swagger.json
{
  "swagger" : "2.0",
  "info" : {
...
    "version" : "1.1.0",
    "title" : "DMaaP Bus Controller REST API",
    "termsOfService" : "http://www.apache.org/licenses/LICENSE-2.0",
    "contact" : {
      "url" : "http://www.onap.org"
    },
    "license" : {
      "name" : "Licensed under the Apache License, Version 2.0",
      "url" : "http://www.apache.org/licenses/LICENSE-2.0"
    }
  },
  "host" : "www.[host]:[port]",
  "basePath" : "/webapi",
...
  "schemes" : [ "http", "https" ],
  "paths" : {
...
    "/mr_clusters" : {
      "get" : {
        "tags" : [ "MR_Clusters" ],
        "summary" : "return MR_Cluster details",
        "description" : "Returns array of  `MR_Cluster` objects.",
        "operationId" : "getMr_Clusters",
        "consumes" : [ "application/json" ],
        "produces" : [ "application/json" ],
        "responses" : {
          "200" : {
            "description" : "Success",
            "schema" : {
              "$ref" : "#/definitions/MR_Cluster"
            }
          },
          "400" : {
            "description" : "Error",
            "schema" : {
              "$ref" : "#/definitions/ApiError"
            }
          }
        }
      },
      "post" : {
        "tags" : [ "MR_Clusters" ],
        "summary" : "return MR_Cluster details",
        "description" : "Create an  `MR_Cluster` object.",
        "operationId" : "addMr_Cluster",
        "consumes" : [ "application/json" ],
        "produces" : [ "application/json" ],
        "responses" : {
          "200" : {
            "description" : "Success",
            "schema" : {
              "$ref" : "#/definitions/MR_Cluster"
            }
          },
          "400" : {
            "description" : "Error",
            "schema" : {
              "$ref" : "#/definitions/ApiError"
            }
          }
        }
      }
    },
    "/mr_clusters/{clusterId}" : {
      "get" : {
        "tags" : [ "MR_Clusters" ],
        "summary" : "return MR_Cluster details",
        "description" : "Retrieve an  `MR_Cluster` object, specified by clusterId.",
        "operationId" : "getMR_Cluster",
        "consumes" : [ "application/json" ],
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "clusterId",
          "in" : "path",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "Success",
            "schema" : {
              "$ref" : "#/definitions/MR_Cluster"
            }
          },
          "400" : {
            "description" : "Error",
            "schema" : {
              "$ref" : "#/definitions/ApiError"
            }
          }
        }
      },
      "put" : {
        "tags" : [ "MR_Clusters" ],
        "summary" : "return MR_Cluster details",
        "description" : "Update an  `MR_Cluster` object, specified by clusterId.",
        "operationId" : "updateMr_Cluster",
        "consumes" : [ "application/json" ],
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "clusterId",
          "in" : "path",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "Success",
            "schema" : {
              "$ref" : "#/definitions/MR_Cluster"
            }
          },
          "400" : {
            "description" : "Error",
            "schema" : {
              "$ref" : "#/definitions/ApiError"
            }
          }
        }
      },
      "delete" : {
        "tags" : [ "MR_Clusters" ],
        "summary" : "return MR_Cluster details",
        "description" : "Delete an  `MR_Cluster` object, specified by clusterId.",
        "operationId" : "deleteMr_Cluster",
        "consumes" : [ "application/json" ],
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "clusterId",
          "in" : "path",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/MR_Cluster"
            }
          },
          "204" : {
            "description" : "Success",
            "schema" : {
              "$ref" : "#/definitions/MR_Cluster"
            }
          },
          "400" : {
            "description" : "Error",
            "schema" : {
              "$ref" : "#/definitions/ApiError"
            }
          }
        }
      }
    },
...
  },
  "definitions" : {
    "ApiError" : {
      "type" : "object",
      "properties" : {
        "code" : {
          "type" : "integer",
          "format" : "int32"
        },
        "message" : {
          "type" : "string"
        },
        "fields" : {
          "type" : "string"
        },
        "is2xx" : {
          "type" : "boolean"
        }
      }
    },
...
    "MR_Cluster" : {
      "type" : "object",
      "properties" : {
        "lastMod" : {
          "type" : "string",
          "format" : "date-time"
        },
        "status" : {
          "type" : "string",
          "enum" : [ "EMPTY", "NEW", "STAGED", "VALID", "INVALID", "DELETED" ]
        },
        "dcaeLocationName" : {
          "type" : "string"
        },
        "fqdn" : {
          "type" : "string"
        },
        "topicProtocol" : {
          "type" : "string"
        },
        "topicPort" : {
          "type" : "string"
        },
        "replicationGroup" : {
          "type" : "string"
        },
        "sourceReplicationPort" : {
          "type" : "string"
        },
        "targetReplicationPort" : {
          "type" : "string"
        }
      }
    },
...
  }
}